server directory refactor. move /login move devServer to server directory move login to root, update login middleware + tests move api utils to new folder close #106
