+++ test/server/REST/clone.spec.js
@@ -1,9 +1,9 @@
 import { expect } from 'chai';
 import { Block as exampleBlock } from '../../schemas/_examples';
-import { set as dbSet } from '../../../server/database';
+import { set as dbSet } from '../../../server/utils/database';
 import request from 'supertest';
 
-const devServer = require('../../../devServer');
+const devServer = require('../../../server/devServer');
 
 describe('REST', () => {
   let server;
