+++ test/server/REST/history.spec.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 import { Project as exampleProject } from '../../schemas/_examples';
-import { set as dbSet } from '../../../server/database';
+import { set as dbSet } from '../../../server/utils/database';
 import request from 'supertest';
-import devServer from '../../../devServer';
+import devServer from '../../../server/devServer';
 
 describe('REST', () => {
   let server;
