+++ test/server/REST/project.spec.js
@@ -3,9 +3,9 @@
 import { expect } from 'chai';
 import { Block as exampleBlock } from '../../schemas/_examples';
 import request from 'supertest';
-import { set as dbSet } from '../../../server/database';
+import { set as dbSet } from '../../../server/utils/database';
 
-const devServer = require('../../../devServer');
+const devServer = require('../../../server/devServer');
 
 describe('REST', () => {
   let server;
