+++ test/server/api/history.spec.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 import uuid from 'node-uuid';
 import { Block as exampleBlock } from '../../schemas/_examples';
-import { getSafe as dbGetSafe } from '../../../server/database';
-import { createDescendant, record, makeHistoryKey, getAncestors, getDescendants, getRoot, getDescendantsRecursively } from '../../../server/history';
+import { getSafe as dbGetSafe } from '../../../server/utils/database';
+import { createDescendant, record, makeHistoryKey, getAncestors, getDescendants, getRoot, getDescendantsRecursively } from '../../../server/utils/history';
 
 describe('History', () => {
   it('makeHistoryKey() should append "-history"', () => {
